refactor: move Checkbox into SharedComponents
