+++ src/components/DownloadAndShareComponents/TermsAndConditionsCheckbox/TermsAndConditionsCheckbox.tsx
@@ -1,10 +1,10 @@
 import { FC } from "react";
 
 import BrushBorders from "@/components/SpriteSheet/BrushSvgs/BrushBorders";
-import Checkbox from "@/components/Checkbox";
+import Checkbox from "@/components/SharedComponents/Checkbox";
 import Box from "@/components/SharedComponents/Box";
 import FieldError from "@/components/FormFields/FieldError";
-import { CheckboxProps } from "@/components/Checkbox/Checkbox";
+import { CheckboxProps } from "@/components/SharedComponents/Checkbox/Checkbox";
 
 export type TermsAndConditionsCheckboxProps = CheckboxProps & {
   errorMessage?: string;
