+++ src/components/SharedComponents/Checkbox/Checkbox.test.tsx
@@ -1,9 +1,10 @@
 import { screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 
-import renderWithTheme from "../../__tests__/__helpers__/renderWithTheme";
-
 import Checkbox from "./Checkbox";
+
+import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
+
 
 describe("Checkbox", () => {
   it("renders a checkbox with a label text", () => {
