+++ src/styles/theme/types.ts
@@ -2,7 +2,7 @@ import { ThemedStyledProps, DefaultTheme } from "styled-components";
 import { z } from "zod";
 
 import { BadgeConfig } from "../../components/Badge";
-import { CheckboxConfig } from "../../components/Checkbox";
+import { CheckboxConfig } from "../../components/SharedComponents/Checkbox";
 import { SelectListBoxConfig } from "../../components/DropdownSelect/ListBox";
 import { MenuConfig } from "../../components/Menu/Menu";
 import { ToggleStyleConfig } from "../../components/Toggle/Toggle";
